+++ src/main/java/es/in2/vcverifier/component/CryptoComponent.java
@@ -5,12 +5,8 @@ import com.nimbusds.jose.jwk.Curve;
 import com.nimbusds.jose.jwk.ECKey;
 import com.nimbusds.jose.jwk.KeyUse;
 import es.in2.vcverifier.config.BackendConfig;
-import es.in2.vcverifier.exception.DidKeyCreationException;
 import es.in2.vcverifier.exception.ECKeyCreationException;
-import es.in2.vcverifier.util.UVarInt;
 import lombok.RequiredArgsConstructor;
-import org.bitcoinj.base.Base58;
-import org.bouncycastle.jcajce.provider.asymmetric.ec.BCECPublicKey;
 import org.bouncycastle.jce.ECNamedCurveTable;
 import org.bouncycastle.jce.spec.ECParameterSpec;
 import org.bouncycastle.jce.spec.ECPrivateKeySpec;
@@ -22,7 +18,6 @@ import java.math.BigInteger;
 import java.security.KeyFactory;
 import java.security.interfaces.ECPrivateKey;
 import java.security.interfaces.ECPublicKey;
-import java.security.spec.X509EncodedKeySpec;
 
 @Configuration
 @RequiredArgsConstructor
@@ -57,53 +52,11 @@ public class CryptoComponent {
             // Build the ECKey using secp256r1 curve (P-256)
             return new ECKey.Builder(Curve.P_256, publicKey)
                     .privateKey(privateKey)
-                    .keyID(generateDidKey(publicKey))
+                    .keyID(backendConfig.getDidKey())
                     .keyUse(KeyUse.SIGNATURE)
                     .build();
         } catch (Exception e) {
             throw new ECKeyCreationException("Error creating JWK source for secp256r1: " + e);
         }
-    }
-
-
-    // Generates a standard DID Key
-    private String generateDidKey(ECPublicKey ecPublicKey) {
-        try {
-            // Convert ECPublicKey (Java) to BCECPublicKey (Bouncy Castle)
-            byte[] encodedKey = ecPublicKey.getEncoded();
-            KeyFactory keyFactory = KeyFactory.getInstance("EC", BouncyCastleProviderSingleton.getInstance());
-            // Decode the public key using Bouncy Castle
-            X509EncodedKeySpec keySpec = new X509EncodedKeySpec(encodedKey);
-            BCECPublicKey bcPublicKey = (BCECPublicKey) keyFactory.generatePublic(keySpec);
-            // Extract the bytes from BCECPublicKey
-            byte[] pubKeyBytes = bcPublicKey.getQ().getEncoded(true);
-            // Multicodec key code for secp256r1
-            int multiCodecKeyCodeForSecp256r1 = 0x1200;
-            // Convert the public key to multibase58 format
-            String multiBase58Btc = convertRawKeyToMultiBase58Btc(pubKeyBytes, multiCodecKeyCodeForSecp256r1);
-            return "did:key:z" + multiBase58Btc;
-        } catch (Exception e) {
-            throw new DidKeyCreationException("Error converting public key to did:key", e);
-        }
-    }
-
-    // Converts raw public key bytes into a multibase58 string
-    private String convertRawKeyToMultiBase58Btc(byte[] publicKey, int code) {
-        UVarInt codeVarInt = new UVarInt(code);
-
-        // Calculate the total length of the resulting byte array
-        int totalLength = publicKey.length + codeVarInt.getLength();
-
-        // Create a byte array to hold the multicodec and raw key
-        byte[] multicodecAndRawKey = new byte[totalLength];
-
-        // Copy the UVarInt bytes to the beginning of the byte array
-        System.arraycopy(codeVarInt.getBytes(), 0, multicodecAndRawKey, 0, codeVarInt.getLength());
-
-        // Copy the raw public key bytes after the UVarInt bytes
-        System.arraycopy(publicKey, 0, multicodecAndRawKey, codeVarInt.getLength(), publicKey.length);
-
-        // Encode the combined byte array to Base58
-        return Base58.encode(multicodecAndRawKey);
     }
 }