+++ src/main/java/es/in2/vcverifier/config/BackendConfig.java
@@ -24,6 +24,10 @@ public class BackendConfig {
         return privateKey;
     }
 
+    public String getDidKey() {
+        return properties.identity().didKey();
+    }
+
     private BackendProperties.TrustFramework getSelectedTrustFramework() {
         return properties.getDOMETrustFrameworkByName();
     }